WIP: just added compare interrupt

+++ src/bus/audio_interface/AudioInterface.hpp
@@ -4,7 +4,7 @@ class AudioInterface {
 public:
     uint32_t dramAddress = 0;
     uint32_t audioLength = 0;
-    uint16_t dacRate = 0;
-    uint8_t bitRate = 0;
+    uint32_t dacRate = 0;
+    uint32_t bitRate = 0;
     bool dmaEnable = false;
 };
