+++ src/bus/peripheral_interface/PeripheralInterface.hpp
@@ -4,17 +4,17 @@
 class PeripheralInterface {
 public:
     PIStatus piStatus;
-    uint8_t dom1Latch = 0;
-    uint8_t dom2Latch = 0;
+    uint32_t dom1Latch = 0;
+    uint32_t dom2Latch = 0;
 
-    uint8_t dom1Pwd = 0;
-    uint8_t dom2Pwd = 0;
+    uint32_t dom1Pwd = 0;
+    uint32_t dom2Pwd = 0;
 
-    uint8_t dom1Pgs = 0;
-    uint8_t dom2Pgs = 0;
+    uint32_t dom1Pgs = 0;
+    uint32_t dom2Pgs = 0;
 
-    uint8_t dom1Rls = 0;
-    uint8_t dom2Rls = 0;
+    uint32_t dom1Rls = 0;
+    uint32_t dom2Rls = 0;
 
     uint32_t dramAddress = 0;
     uint32_t cartAddress = 0;
